backend/requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
google-generativeai
python-dotenv
